bot/dev/requirements.txt
aiogram==3.7.0
python-dotenv
audio-extract
fastapi
supabase
aiofiles
pydub
requests
docx
python-docx
docx2pdf